+++ src/main/java/gift/Application.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
-import gift.service.JwtAuthFilter;
+import gift.config.JwtAuthFilter;
 
 @SpringBootApplication
 public class Application {
